requirements.txt
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.0
pandas==2.2.3
gunicorn==23.0.0
whitenoise==6.9.0
requests==2.31.0
apscheduler==3.10.4
numpy==1.26.4
scikit-learn==1.4.2
pytz==2024.1
SQLAlchemy==2.0.28